backend/requirements.txt
fastapi[standard]>=0.136.0
sqlalchemy>=2.0
alembic>=1.13
openai>=1.0
python-dotenv>=1.0
pandas>=2.0
pydantic>=2.0
httpx>=0.27
PyJWT>=2.8
oss2>=2.18
pytest>=8.0
